add stdmin.lua
