+++ src/logic/scripting/lua/lua_engine.hpp
@@ -7,13 +7,15 @@
 #include "logic/scripting/scripting_functional.hpp"
 #include "lua_util.hpp"
 
+class EnginePaths;
+
 namespace lua {
     enum class StateType {
         BASE,
         GENERATOR,
     };
 
-    void initialize();
+    void initialize(const EnginePaths& paths);
     void finalize();
 
     bool emit_event(
@@ -22,7 +24,7 @@ namespace lua {
         std::function<int(State*)> args = [](auto*) { return 0; }
     );
     State* get_main_state();
-    State* create_state(StateType stateType);
+    State* create_state(const EnginePaths& paths, StateType stateType);
     [[nodiscard]] scriptenv create_environment(State* L);
 
     void init_state(State* L, StateType stateType);
